+++ OWMS/jis_job_backend/src/main.ts
@@ -111,6 +111,7 @@ async function bootstrap() {
     origin: [
       'http://localhost:3000',
       'http://localhost:1420',
+      'http://192.168.123.75:3000',
       'http://192.168.123.46:3000',
       'tauri://localhost',
     ], // 프론트엔드 및 Tauri 앱 허용